requirements.txt
fastapi==0.100.0
uvicorn[standard]==0.23.0
motor==3.2.0
pymongo==4.4.1
python-dotenv==1.0.0
pydantic[email]==2.0.3
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
